now used consumables, their effects and durations are properly stored in the database
changed some of the consumables' fields from int to double, fixed player not taking reduced damage because of defense, fixed minor bugs and changed effect related classes

+++ Agoraphobia/AgoraphobiaAPI/Dtos/Consumable/ConsumableDto.cs
@@ -7,10 +7,10 @@ public class ConsumableDto
     public string Description { get; set; } = string.Empty;
     public int RarityIdx { get; set; }
     public int Price { get; set; }
-    public int Sanity { get; set; }
+    public double Sanity { get; set; }
     public int Energy { get; set; }
-    public int Hp { get; set; }
-    public int Defense { get; set; }
-    public int Attack { get; set; }
+    public double Hp { get; set; }
+    public double Defense { get; set; }
+    public double Attack { get; set; }
     public int Duration { get; set; }
 }
